refactor: 로또 개수 메서드  파라미터 객체화
