+++ src/main/java/lotto/domain/LottoNumber.java
@@ -2,6 +2,9 @@ package lotto.domain;
 
 public class LottoNumber {
 
+    private static final int LOTTO_VALID_NUM_START = 1;
+    private static final int LOTTO_VALID_NUM_END = 45;
+
     private final int lottoNum;
 
     public LottoNumber(int lottoNum) {
@@ -11,7 +14,7 @@ public class LottoNumber {
     }
 
     private void validateNumberRange(final int number) {
-        if (number < 1 || number > 45) {
+        if (number < LOTTO_VALID_NUM_START || number > LOTTO_VALID_NUM_END) {
             throw new IllegalArgumentException("복권의 숫자 범위가 유효하지 않습니다.");
         }
     }
