+++ src/test/java/lotto/LottoCalculatorTest.java
@@ -4,8 +4,10 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 import lotto.domain.LottoCountResult;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoNumbers;
+import lotto.service.LottoCountCalculator;
 import org.junit.jupiter.api.Test;
 
 class LottoCalculatorTest {
@@ -14,9 +16,9 @@ class LottoCalculatorTest {
     void 당첨_번호와_몇개가_일치하는지_판단할_수_있다() {
         // given
         LottoCountCalculator calculator = LottoCountCalculator.getInstance();
-        List<Integer> lotteryNumbers = new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 6));
-        int bounusNum = 6;
-        List<Integer> userNumbers = new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 8));
+        LottoNumbers lotteryNumbers = new LottoNumbers(new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 6)));
+        LottoNumber bounusNum = new LottoNumber(6);
+        LottoNumbers userNumbers = new LottoNumbers(new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 8)));
 
         // when
         LottoCountResult lottoResult = calculator.countLotteryNumber(lotteryNumbers, userNumbers, bounusNum);
